chore: move createTarget function to a separate file in utils folder
